mudanças em páginas, footer, header e imagens de fundo
