+++ my-app/.vscode/settings.json
@@ -1,0 +1,15 @@
+{
+  "cSpell.words": [
+    "card",
+    "Card",
+    "cards",
+    "dream",
+    "experience",
+    "Pokedex",
+    "pokemon",
+    "Pokemon",
+    "pokemons",
+    "Pokemons",
+    "sprites"
+  ]
+}
