+++ my-app/src/App.css
@@ -1,21 +1,21 @@
 .app {
     min-height: 100vh;
-
 }
 
-
-/* testando buttons e inputs */
 form {
     max-width: 600px;
     margin: 0 auto;
     padding: 60px;
-   
 }
 
 input, select, textarea, .radios, .checks {
     margin-bottom: 20px;
-
 }
+
 .radios label, .checks label {
-margin-right: 20px;
+    margin-right: 20px;
+}
+
+button:active {
+    transform: translate(0em, 0.2em);
 }
