+++ my-app/src/Components/Card/index.css
@@ -1,7 +1,6 @@
-.iconCard {
-  height: 30px;
+.cardModel {
+  height: 100vh;
   display: flex;
-
 }
 
 .image img {
@@ -11,6 +10,7 @@
   min-height: 170px;
   display: flex;
   align-content: center;
+  margin-bottom: 10px;  
 }
 
 .weight, .height, .power {
@@ -20,23 +20,19 @@
 .id {
   width: 100%;
   font-family: 'Carter One', Times, serif;
-  font-size: 40px;
+  font-size: 18px;
   display: flex;
-  color: #cccccc;
+  color: white;
   padding: 0 20px;
 }
 
 .nameCard {
   font-family: 'Seymour One', sans-serif;
-  background-color: #008584;
-  color: #cccccc;
+  background-color:rgb(11,74,77);
+  color: white;
   border-radius: 50px;
 }
 
 .positionNameCard {
   width: 100%;
-}
-
-.cardModel {
-  background-color: red;
 }
