--- my-app/src/Components/Register/index.css
@@ -1,39 +1,0 @@
-.register {
-    width: 100%;
-    height: 920px;
-    background: url('../../Assets/Images/Register.jpg');
-    background-size: cover;
-    background-color: #51647a;
-
-
-}
-
-.register input {
-    border-radius: 10px;
-    margin-top: 10px;
-}
-
-.register p {
-    font-family: 'Roboto', sans-serif, Times;
-    font-size: 25px;
-    color: white;
-    margin-left: 10px;
-}
-
-.formData form {
-    height: 500;
-    display: flex;
-    flex-direction: column;
-    margin-left: 40px;
-    border-radius: 20px;
-    background-color: rgba(0, 0, 0, 0.3);
-    margin-top: 30px;
-}
-.formData  {
-
-    min-height: 720px;
-padding-right: 30px;
-
-
-} 
-
