+++ my-app/src/Pages/Connect/index.css
@@ -1,24 +1,7 @@
 .pageConnect {
     display: flex;
     flex-direction: column;
-    align-items: center;
-    align-items: flex-end;
-    min-height: 100vh;
-    background-image: url('../../Assets/Images/backgroundMain.jpg');
-    background-repeat: no-repeat;
-    background-repeat: no-repeat, repeat-x;
-    background-position: top right, center;
-    background-size: cover;
-}
-
-.pageConnect form {
-    height: 300px;
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    border-radius: 20px;
-    background-color: rgba(0, 0, 0, 0.3);
-    margin: 30px 40px 0 0;
+    width: 100%;
 }
 
 .inputConnect {
@@ -67,7 +50,7 @@
     bottom: 0;
     width: 100%;
     height: 2px;
-    background:  rgb(0, 201, 169);
+    background: rgb(0, 201, 169);
     border-radius: 4px;
     transition: 0.5s;
     pointer-events: none;
@@ -123,7 +106,7 @@
     bottom: 0;
     width: 100%;
     height: 2px;
-    background:  rgb(0, 201, 169);
+    background: rgb(0, 201, 169);
     border-radius: 4px;
     transition: 0.5s;
     pointer-events: none;
@@ -136,7 +119,34 @@
 }
 
 .formContainer {
-    min-height: 720px;
+    background-image: url('../../Assets/Images/ConnectBackground.jpg');
+    background-repeat: no-repeat;
+    background-position: top right, center;
+    background-size: cover;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
+.formContainer form {
+    display: flex;
+    flex-direction: column;
+    height: 300px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    align-items: flex-end;
+    align-self: flex-end;
+    border-radius: 20px;
+    background-color: rgba(0, 0, 0, 0.3);
+    margin-top: 30px;
+}
+
+.connect {
+    display: flex;
+    flex-direction: column;
+    align-self: flex-end;
+    margin-right: 40px;
 }
 
 .buttonsHome {
@@ -153,9 +163,9 @@
     background: transparent;
     text-transform: uppercase;
     font-weight: 900;
-  }
-  
-  .buttonsHome:before {
+}
+
+.buttonsHome:before {
     position: absolute;
     content: '';
     left: 0;
@@ -166,9 +176,9 @@
     border-left: 4px solid transparent;
     box-sizing: border-box;
     transform: translateX(100%);
-  }
-  
-  .buttonsHome:after {
+}
+
+.buttonsHome:after {
     position: absolute;
     content: '';
     top: 0;
@@ -179,31 +189,28 @@
     border-right: 4px solid transparent;
     box-sizing: border-box;
     transform: translateX(-100%);
-  }
-  
-  .buttonsHome:hover {
+}
+
+.buttonsHome:hover {
     box-shadow: 0 5px 15px rgba(0, 0, 0, 0.5);
-  }
-  
-  .buttonsHome:hover:before {
-    border-color:  rgb(0, 201, 169);
+}
+
+.buttonsHome:hover:before {
+    border-color: rgb(0, 201, 169);
     height: 100%;
     transform: translateX(0);
     transition: .3s transform linear, .3s height linear .3s;
-  }
-  
-  .buttonsHome:hover:after {
-    border-color:  rgb(0, 201, 169);
+}
+
+.buttonsHome:hover:after {
+    border-color: rgb(0, 201, 169);
     height: 100%;
     transform: translateX(0);
     transition: .3s transform linear, .3s height linear .5s;
-  }
-  
-  button {
-    color: white;
-    text-decoration: none;
-    cursor: pointer;
-    outline: none;
-    border: none;
-    background: transparent;
-  }
+}
+
+.buttonSubmit {
+    margin-top: 15px;
+    max-height: 30px;
+    max-width: 200px;
+}
