+++ my-app/src/Pages/HomePage/index.js
@@ -13,7 +13,6 @@ const HomePage = () => {
       <Header/>
       <Main/>
       <Footer/>
-      {/* <CursoCss/> */}
     </div>
   );
 };
