+++ my-app/src/Pages/Register/index.css
@@ -1,0 +1,44 @@
+.register {
+    width: 100%;
+    /* height: 920px; */
+    /* min-height: 100vh; */
+    display: flex;
+    flex-direction: column;
+ 
+}
+
+.register input {
+    border-radius: 10px;
+    margin-top: 10px;
+}
+
+.register p {
+    font-family: 'Roboto', sans-serif, Times;
+    font-size: 25px;
+    color: white;
+    margin-left: 10px;
+}
+
+.formData form {
+    height: 500;
+    display: flex;
+    flex-direction: column;
+    margin-left: 40px;
+    border-radius: 20px;
+    background-color: rgba(0, 0, 0, 0.3);
+    margin-top: 30px;
+   
+}
+
+.formData {
+    min-height: 100vh;
+    /* min-height: 720px; */
+    padding-right: 30px;
+    background: url('../../Assets/Images/RegisterBackground.jpg');
+    background-size: contain;
+    background-repeat: no-repeat;
+    background-repeat: no-repeat, repeat-x;
+    background-position: top right, center;
+    background-size: cover;
+
+}
